Small adjustments on database (models, factories, etc.)

+++ database/migrations/2023_03_17_204917_create_expenses_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 ->constrained('users')
                 ->cascadeOnDelete();
             $table->string('description');
-            $table->bigInteger('amount');
+            $table->decimal('amount', 10, 2);
             $table->timestamps();
         });
     }
